requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# tests/linter
black
coverage
flake8
pytest
yapf

# data science
numpy
scikit-learn
Pillow
